apply suggestions and move sector_alignment
